Add Bootcamp and Course Ownership

+++ controllers/bootcampsC.js
@@ -45,7 +45,19 @@ exports.getBootCamp = asyncHandler(async (req, res, next) => {
 // @route     POST /api/v1/bootcamps
 // @access    Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+  // add user to req.body
+  req.body.user = req.user.id;
+
+  // Check for published bootcamp
+  const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+  // if user is not an admin, they can add only one bootcamp 
+  if (publishedBootcamp && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`The user with ID ${req.user.id} has already published bootcamp`, 400))
+  }
+
   const bootcamp = await Bootcamp.create(req.body);
+
   res.status(201).json({ success: true, data: bootcamp })
 });
 
@@ -56,11 +68,20 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route     PUT /api/v1/bootcamps/:id
 // @access    Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+  let bootcamp = await Bootcamp.findById(req.params.id);
 
   if (!bootcamp) {
     return next(new ErrorResponse(`Bootcamp not found with id ${req.params.id}`), 404);
   }
+
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401))
+  }
+
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+
   res.status(200).json({ success: true, data: bootcamp })
 });
 
@@ -73,7 +94,12 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
   const bootcamp = await Bootcamp.findById(req.params.id);
 
   if (!bootcamp) {
-    return next(new ErrorResponse(`Bootcamp not found with id ${req.params.id}`), 404);
+    return next(new ErrorResponse(`Bootcamp not found with id ${req.params.id}`, 404));
+  }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401))
   }
 
   await bootcamp.deleteOne();
@@ -122,6 +148,12 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   if (!bootcamp) {
     return next(new ErrorResponse(`Bootcamp not found with the id of ${req.params.id}`, 400));
   }
+  
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401))
+  }
+
   // console.log(req.files);
 
   if (!req.files) {
