+++ controllers/coursesC.js
@@ -50,10 +50,16 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
   if (!bootcamp) {
     return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`, 404));
+  }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`, 401))
   }
 
   const course = await Course.create(req.body);
@@ -71,7 +77,12 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
 exports.updateCourse = asyncHandler(async (req, res, next) => {
   let course = await Course.findById(req.params.id);
   if (!course) {
-    return next(new ErrorResponse(`No bootcamp with the id of ${req.params.id}`, 404));
+    return next(new ErrorResponse(`No course with the id of ${req.params.id}`, 404));
+  }
+
+  // Make sure user is bootcamp owner
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${course._id}`, 401))
   }
 
   const preBootCampId = course.bootcamp;
@@ -106,7 +117,12 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
   const course = await Course.findById(req.params.id);
   if (!course) {
-    return next(new ErrorResponse(`No bootcamp with the id of ${req.params.id}`, 404));
+    return next(new ErrorResponse(`No course with the id of ${req.params.id}`, 404));
+  }
+
+  // Make sure user is bootcamp owner
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${course._id}`, 401))
   }
 
   await course.deleteOne();
